sentinel/pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.13.3",
    "discord.py==2.7.1",
    "pydantic>=2.0.0,<3.0.0",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "RUF", "SIM", "UP", "W"]
ignore = ["E501", "E203", "E231"]
